Fix vrrp instances going to fault state when have virtual routes
If an interface going down caused a vrrp instance to go to fault
state, and the vrrp instance also had virtual routes, the state
of the vrrp instance would be set to backup when the deletion of
the virtual route was detected. This commit ensures that the vrrp
instance stays in fault state until the interface is brought up
again.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>