Remove Red Hat Linux 9 and RH Enterprise Linux 3 from spec file
Red Hat Linux 9 and Red Hat Enterprise Linux 3 are both based on
Linux 2.4, which is no longer supported by keepalived. The options
in the spec file for Reh Hat Linux 9 have twice caused people to
specify wrong options to configure when trying to build keepalived,
so the options are removed to i) avoid confusion and ii) they are
not longer relevant.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>