Ensure all protocol headers received before return pointer to vrrp header
vrrp_get_header() returns a pointer to the vrrp header, but it now returns
NULL if insufficient data has been received to include all the (IP,
possibly AH, and VRRP) headers (this does not include the VIPs in the VRRP
packet).

This means that when a pointer to the VRRP header is returned, all fields in
all protocol headers can safely be accessed.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>