Enhance and streamline checking of validity of received VRRP packet
This includes checking that a packet is multicast, unless unicast is
expected in which case it is checked for unicast, ensuring that if
AH authentication is used, the next header protocol is VRRP.

The sequence of some checks is revised to ensure that the fields being
checked are valid to be accessed prior to accessing them, e.g. check
that the packet is VRRP version 2 before checking the authentication.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>