Handle interface MAC addresses changing
When an interface is added to a bond interface, if it is the first
interface added, the MAC address of the bond interface is changed
to the MAC address of the added interface. When subsequent interfaces
are added, their MAC addresses are changed to that of the bond
interface.

Issue #1112 identified that if a bond interface is deleted and
recreated, the gratuitous ARPs were sent with the wrong source MAC
address.

This commit now updates interface MAC addresses from the netlink
RTM_NEWLINK messages, so that the correct MAC address is always
used.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>