Fix building on pre 3.10 kernels re track_process
Issue #1119 reported that keepalived wouldn't build on CentOS 6.

Various PROC_EVENT_* declarations were assumed to exist, some of which
were not introduced until Linux v3.10. Most of them are not needed, but
PROC_EVENT_COMM is used by the track_process code.

This commit now checks for the existence of the PROC_EVENT_* declarations,
but since keepalived uses PROC_EVENT_COMM, track_process is not supported
prior to Linux v3.2.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>