Make track_process work prior to Linux 3.2, but with limitations
Prior to Linux 3.2 the PROC_EVENT_COMM event did not exist, which
means that keepalived is unable to detect changes to process name
(/proc/PID/comm) prior to Linux 3.2. most processes do not change
their process name, and so using track_process prior to Linux 3.2
is safe so long as the monitored processes are known not to change
their process name.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>