Don't remove link local IPv6 address from VMAC that isn't keepalived's
If IFLA_INET6_ADDR_GEN_MODE isn't supported and a macvlan interface
already had a (non-default) link local addresss and the link local
address that matched the interface's MAC address was added, keepalived
was removing it as soon as it was added. This commit stop keepalived
removing the address when we shouldn't.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>