Fix handling of VMACs with multiple reloads
If a configuration is loaded that has a VRRP instance using a VMAC,
then the configuration is updated to remove that VRRP instance and
keepalived reloads its configuration, then the configuration is
updated again to reinstate the VRRP instance and the configuration
is again reloaded, keepalived thought the VMAC interface still
existed, whereas it was deleted following the first reload.

This commit ensures that keepalived properly detects whether an
interface exists following a reload.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>