Remember more than one interface local address per interface
Keepalived needs a local address for each interface it sends adverts
on. If the address keepalived is using is deleted and another address
is configured on the interface, then keepalived should start using
that address. To do this, a list of configured address on each
interfaces needs to be maintained.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>