Don't consider VIPs as local addresses when restart after crash
Keepalived maintains a list of addresses per interface that can be
used as source adddresses for adverts. To build the list, keepalived
reads the addresses configured on interfaces when it starts. However,
if keepalived crashed it will have left VIPs configured on interfaces,
and we don't want to use them as advert source addresses.

This commit makes keepalived compare the addresses on interfaces
to VIPs, and ignores any addresses that are VIPs.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>