Allow fractional warmup, delay_loop and delay_before_retry for checkers
To shorten the real server monitoring interval, make it possible to specify
decimal value for following items:
  warmup
  delay_loop
  delay_before_retry

This patch is based on a patch submitted by tamu.0.0.tamu@gmail.com. The
only changes are to add warmup, ensure all places where these options can be
specified are updated, and to report the setting with SNMP and in the config
dump (connect_timeout will be handled in a subsequent patch).

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>