Add option not to check for EINTR if using signalfd()
If keepalived is using signalfd(), there are no asynchronous signal
handlers, and therefore EINTR cannot be returned.

Currently the check for EINTR is enabled by default, and configure
option --disable-eintr-debug disables the check, while
--enable-eintr-debug enables writing log entries if EINTR is returned.

Once sufficient testing has been performed, the default will be
changed not to test for EINTR if signalfd() is supported.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>