Fix log output when real server removed
FMT_VS() and FMT_RS() both call inet_sockaddrtotrio which uses a
static buffer to return the formatted string, but since FMT_VS(),
wheich simply calls format_vs() copies the returned string to its
own static buffer, if FMT_VS() was called before FMT_RS() then
the returned strings from both could be used.

The problem occurs when both FMT_VS() and FMT_RS() are used as
parameters to log_message() (or printf etc). It appeared to work
fine on x86_64, but was writing the same IP address for both the
real server and virtual server on ARM architectures. This is due
to the compiler evaluating parameters to the log_message() function
call in a different order on the different architectures.

This commit adds inet_sockaddrtotrio_r() which allows the output
to be in a buffer specified by the caller, and so FMT_VS() and
FMT_RS() can now be called in either order without one overwriting
a buffer used by the other.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>