Fix sending notifies for vrrp instances at startup when in sync group
Issue #1155 idenfified that notify scripts for vrrp instance transition
to backup state when keepalived started up were not being sent if
the vrrp instance was in a sync group. It was also the case that SNMP
traps, SMTP alerts and FIFO notifies were not being sent either.

This commit make keepalived send the initial notifies when the vrrp
instance is in a sync group.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>