Create separate checker for each host in SMTP_CHECK block
Having multiple host entries in an SMTP_CHECK block is deprecated.
This commit streamlines the SMTP_CHECK code by creating a separate
SMTP checker for each host declared in the SMTP_CHECK block, so that
apart from parsing the configuration, the code no longer handles
multiple hosts per checker.

The support for parsing configuration with multiple hosts is only
enabled if WITH_HOST_ENTRIES is defined in check_smtp.c. It is
currently enabled, but when support for multiple hosts in the
SMTP_CHECK block is finally removed, it will simply be a matter of
deleting all code in the WITH_HOST_ENTRIES conditional blocks.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>