Don't set SO_LINGER with a timeout of 0
SO_LINGER with a timeout of 0 causes a TCP connection to be reset
rather than cleanly closed. Instead of specifying a timeout of 0,
use 5 seconds, so that there is an orderly shutdown of the TCP
connection, but the close socket doesn't remain in TIMED_WAIT state
for more than a short time.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>