Log warnings if real server and virtual server ports don't match
This commit adds logging warnings if virtual and real server ports,
when using TUN or DR, don't match.

It also sets the real server ports to be the same as the virtual server
ports.  Although listing the IPVS configuration with ipvsadm will look
different, the kernel ignored the port of a real server when using DR
or TUN, so the behaviour isn't changed, but when looking at the
configuration it now shows what is actually happening.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>