Handle macvlans/macvtaps being moved into different namespace from parent
If a macvlan or macvtap interface is moved into a different namespace from
its parent, and the interface is in the namespace in which keepalived is
running, keepalived is unable to get information about, or configure, the
parent interface. In this case, treat the macvlan/macvtap interface as though
it doesn't have a parent interface.

There are a couple of consequences of this in this situation:
1) If a vrrp instance is configured with use_vmac and its configured interface
   is such a macvlan/macvtap interface, keepalived cannot ensure that the
   arp_ignore and arp_filter settings are correct on the parent
2) keepalived cannot check that there a not duplicate VRIDs being used on the
   interface.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>