Fix commit 128bfe6 for pre v4.0 kernels
Commit 128bfe6 - "Ignore base interfaces of macvlans if in a different
namespace" added using IFLA_LINK_NETNSID to detect if the parent of an
interface was in a different namespace. Unfortunately that was only
introduced in Linux v4.0, so don't attempt to use it if it is not
defined.

For kernels older than v4.0 if a macvlan interface's parent is in
another network namespace, but the ifindex of the parent interface also
exists in the namespace in which keepalived is running, then keepalived
will believe the parent of the macvlan is the wrong interface.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>