Fix sending priority 0 adverts after reload for deleted vrrp instances
During a reload, vrrp_dispatcher_release() was called prior to reloading
the configuration, and it closed all the vrrp send/receive sockets. However
it isn't until after the reload that it is known which vrrp instances no
longer exist, and clear_diff_vrrp() attempted to send 0 priority adverts
for those instances. Since the sockets had already been closed, the adverts
could not be sent. Worse, the socket_t structures had been released, but the
released memory was accessed in attempting to send the adverts.

This commit delays calling vrrp_dispatcher_release() until after the new
configuration has been reloaded, and it sends 0 priority adverts before
all the old sockets are closed. Following this new sockets are opened.

It would be possible to make the code more efficient and retain the sockets
that still need to be used, rather than closing them and opening new ones,
but that is for another commit.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>