Ensure virtual servers are properly removed when reloading
Pull request #1246 provided a patch to resolve the issue of virtual
servers in a virtual server group that are deleted from the virtual
server group on a reload weren't being removed from the IPVS
configuration. However, the patch didn't quite work with the current
HEAD of the master branch.

This commit incorporates that patch provided and makes the necessary
adjustments for it to work correctly.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>