Don't enclose /dev/tcp/127.0.0.1/22 in ' chars when running as script
RedHat identified a problem with scripts like:
    vrrp_script {
        script "</dev/tcp/127.0.0.1/22"
    }
where returning an exit code of 127 (script not found).

This was identified to be due to the "script" being enclosed in '
characters, so the resulting system call was
system("'</dev/tcp/127.0.0.1/22'"), which failed. Not adding the leading
and trailing ' characters when the first character of the script is '<'
or '>' resolves the problem.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>