genhash: ask the server to disable keep-alive
genhash speaks only HTTP/1.0 which strictly speaking doesn't support
keep-alive; but some servers which do not support version 1.0 may
interpret the absence of a "Connection" header as "enable keep-alive",
as required by version 1.1 of the protocol.  This breaks genhash,
which waits until the server closes the connection.

To fix this, this patch adds "Connection:close" which explicitly
disables keep-alive.