Ensure that -lm linker library flag is always set
configure was testing whether it was necessary to add the -lm option,
but for some reason gcc adds it itself if -Os is not specified, but
does not add it if -Os is specified. Consequently if configure was
run without -Os, and make was run with -Os the link failed.

The commit ensures that -lm is always used.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>