support http status_code group
The origin status_code only support one specific code, now we can
support http status_code of the same class. That's to say, we can
use 1xx to represent 100-199, 2xx means 200-299 ans so on.

eg: The configure as follows:

	url {
		path /index.html
		status_code 2xx 3xx
	}

which means we consider all status_code range in [200,399] is ok.
Of course the following configure is either 200 or [300,399] is ok.

	url {
		path /index.html
		status_code 2xx 3xx
	}