Use standard bit testing and setting functions
Commit c7c23a2 - "support http status_code group" added additional
bit testing and setting functions, rather than using the already
defined ones in bitops.h.

This commit also resolves the assumption that longs are 64 bits, and
will allow the code to work with longs of any length.

The original commit would cause all status codes 100 to 599 to be
written when the configuration was dumped, regardless of whether
the specific codes were set. This commit now writes the status codes
in ranges.

Finally, if no status code is configured, it sets the bits for the
default status codes (200-299).

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>