Send GARP/NA message when leaving fault state if using unicast
If the master's ARP entry for a backup route has expired and we are
using a short advert interval (< 0.5 seconds), then the backup router
could timeout receiving adverts before the master sends its next
ARP/NDISC message; until it has had a reply to that it cannot send any
adverts to the backup router in question.

This commit makes a VRRP instance that is using unicast send a GARP/NA
when it transitions out of fault state, to ensure that the master (or
local router) can send adverts to us immediately.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>