Fix HTTP_GET with two or more urls
When we config two or more urls,. With the following step
to reproduce:
1. let the two urls check success
2. modify the second url to fail and reload
3. the real server will never be offline

the reason is we will always check the first url after we have
been online. As the reply of issue:
https://github.com/acassen/keepalived/issues/1301#issuecomment-501049906
All URLs should be checked. So when we successfully check one
URL, we should keep the retry_it and check the next URL until we failed,
adding the retry_it and start another round check. Of course,
when we check all URLs successfully, we should reset retry_it
and url_it.

Signed-off-by: Jie Liu <liujie165@huawei.com>