Fix potential segfault when keywords without value
As the pull request
https://github.com/acassen/keepalived/pull/1307#issue-287524770
says: when we meet the situation of like those who call
set_value, but the value is missing. Although we return the NULL
pointer, but it will still meet segfault. We can easily to reproduce
just like the config without a path:

 HTTP_GET {
 	url {
		path
		status_code 200
	}
}

So I consider it necessary to ensure all keywords using set_value
should have a value.

Signed-off-by: Jie Liu <liujie165@huawei.com>