Set thread parameter value explicitly to 0 when add timer thread
It is possible for a function to be called either from a timer thread
or an event thread. When an event thread is added, a vlue can be passed
which will be passed to the function, but currently there is no way to
set the value for a timer thread (a function thread_add_timer_val() can
be added when needed), but in order to allow the value to be used with
an event thread, it needs to be explicitly set to something when called
via a timer thread, so just set it to 0.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>