When an HTTP_GET url check fails, keep checking that URL until success
When a URL check has failed, there is no point checking other URLs until
we know the one that has failed is working again. The approach now is
that the failed URL is checked until it is Ok again, and then all the URLs
are checked before the checker is successful. This will reduce the recovery
time once the failed URL recovers.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>