Make set_value() check for missing parameter
Pull request #1308 identifed that if set_value() was called when
there wasn't a parameter on the command line, keepalived could
segfault since NULL was returned (examples were HTTP_GET with an
empty path specified, and DNS_CHECK with empty name).

This commit modifies set_value() so that keepalived will exit if
it is called with no keyword parameter is missing. Uses of
set_value() where no parameter did not cause a problem (e.g. where
the whole option was optional, such as virtual_host) now check if
the parameter is mising and report a configuration error.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>