Increase open file limit for checker process if no of checkers need it
TCP, HTTP/SSL, DNS and SMTP checkers all use a socket. If there is a
sufficiently large number of checkers, the default open file limit may
be exceeded. This commits counts the number of such checkers, and also
thr number of smtp_alerts, and if necessary increases the open file limit
to allow them all to run at once.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>