Don't send bfd events to vrrp or checker process if no configuration
If there is no vrrp configuration, or no checker configuration, there
is no point sending bfd_event notifications to the relevant processes.
Actually, since the processes may not be running, sending such
notifications can cause the pipes to become full, so it is necessary,
as well as desirable, not to write events to the pipes in those
circumstances.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>