Ensure BFD source port in range 49152..65535
RFC5881 requires the source port for BFD packets to be in the above
range, but keepalived was allowing the port to be randomly generated
by the kernel, and hence could be outside the range.

This commit sets the permitted port range to the intersection of
[49152, 65535] and the values in /proc/sys/net/ipv4/ip_local_port_range,
unless the intersection is too small, in which case it just uses the BFD
specified values.

keepalived generates a random port number in the required range, and then
loops through the range starting from the random port number until it
finds one it can bind to.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>