Make PID files group and world readable
Issue #1378 identified that PID files were created without group
and work read in the file permission bits being set, due to the umask.
This was causing a problem, since a non-root user was needing to read
the PID file.

This commit now forces the file permissions of PID files to be:
owner=rw, group=r, other=r.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>