Revert "Fix route add/delete on reload if only change via address"
This reverts commit f54c2e8294c6f2080c3ae951ba25bf40d5b8d211.

The commit did the wrong thing: the via address is not part of the
key of the route.

The problem was that if it detected it already had a route with the
same key, it replaced the route, but with the old route and not the
new one. The next commit will rectify this.