Correct handling of replacing routes on reload
Issue #1390 identified that a route with a changed source address
wasn't being changed on reload, and identified that the problem was
related to issue #1220.

It turns out that commit f54c2e8 which resolved issue #1220 was not
the correct fix. The problem was that when reloading, if a new route
matched an old route, the old route was replaced with the OLD route
(i.e. it did nothing), rather than replacing it with the NEW route.

This commit now replaces the old route with the new route.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>