Handle changing virtual route to use new VIP on reload
During a reload, routes replacing existing routes are replaced, as
opposed to deleted and added; this avoids the route disappearing for
a short while.  However, if a new route uses a VIP that didn't exist
in the old configuration then the route replacement will fail.

The code is now changed so that it attempts to replace the route, but
if that fails, it deletes the old route and later adds the new route
after the new VIPS have been added.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>