Properly handle MISC_CHECK script returning exit status 0
If a MISC_CHECK script returned an exit status >= 2, and then the
script returns 0, the weight of the real server was not updated
following the script returning 0, and the quorum also wasn't updated.

This commit makes keepalived update the weight of a real server and
the quorum following a MISC_CHECK returning a 0 exit status.

This is a change of behaviour for MISC_CHECKs whose scripts return
an exit status >= 2 subsequently followed by an exit status of 0.
However, the new behaviour is consistent with the documentation and
is also the behaviour that would be expected. Some users may need to
modify their MISC_CHECK scripts if the scripts return an exit code >= 2
and subsequently return an exit code of 0 expecting that to mean that
the status hasn't changed.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>