Handle script timeouts when child process has terminated
Issue #1364 identified that if a track script timed out and the kill
of the process failed to to its exit status already having been reaped,
keepalived would never run the track scripts again. It transpires that
the same problem existed with MISC_CHECK scripts.

The commit now ensures that after the timeout the script will be set
to idle state, so that it can be run again.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>