Enable FIFOs to receive initial fault notifications at startup
Issue #1399 identified that if a track_file caused a VRRP instance
to start in fault state, there was no notification of the fault sent
to the notify FIFOs.

This commit implements the suggestion of chenwng in issue #1399 and
moves the opening of the notify FIFOs to earlier in the init process.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>