Add support for nftables blocking IGMP/MLD packets on VMAC interfaces
Issue #882 identified that VRRP instances using VMACs sent IGMP/MLD
packets using the MAC address of the VMAC interface even when the
vrrp instance was in backup state. This meant that switches would
update what interface the VRRP MAC address was on to the backup
instance, thereby meaning that until the master instance sent another
advert, packets destined to be forwarded by the master instance would
be delivered to the backup instance and lost.

This commit adds support to use nftables to stop the packets being
send from the VMAC interface (and thereby using the 00:00:5e:00:0x:xx
MAC address), and instead the packets are sent on the parent interface.
How this is acheived depends on whether the kernel supports the dup
statement; if it does the IGMP messages are simply moved from the VMAC
interface to the parent interface otherwise the join groups are done on
both the VMAC interface and the parent interface, and nftables drops
the join messages on the VMAC interface.

This functionality might be better implemented using eBPF, but older
systems don't support that, and I still need to work out how to use
eBPF.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>