Don't log an error when deleting nonexistant nftable at startup
Since there is no native flush command to the kernel to delete all
the rules, sets, meters etc in a table, we just delete the table,
and if the table doesn't exist an error is returned.

This commits stops the error being logged, since it isn't an error
of interest.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>