Make skip_check_adv_addr work properly
Keepalived was checking the received advert packet length against
the expected length based on the number of VIPs configured on the
vrrp instance. This commit changes the check so that the received
packet length is checked against the number of VIPs in the VRRP
packet header, thereby ensuring that the advert packet is self
consistent.

The check for the number of addresses is now only done if
skip_check_adr_addr is not set (which matches what the RFC says).
Note: skip_check_adv_addr only skips the check of VIPs in a received
packet if the advert is received from the same master as the previous
advert.

With this commit, if skip_check_adv_addr is set, it is possible to
reload a master instance with a different number of VIPs, and then
subsequently reload the backup instance, without the backup instance
becoming master due to the mismatch of the number of VIPs.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>