Resolve file descriptor errors on reload
If a read or write thread was on the thread reaady queue when a
reload was processed, the file descriptor was not removed from the
epoll instance.

This commit ensures that file descriptors relating to threads on
the thread ready queue are removed from the epoll instance during
a reload.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>