Correct detection of success loading module xt_set
The detection of success loading the module was reversed, so that
if it failed it thought it was successful, and vice versa. This
meant that if it was not successful loading the module, it would
attempt to use ipsets, and if it was successful loading the module
it would not. If the module was already loaded, then there was no
problem.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>