Ensure bfd thread_time_to_wakeup() doesn't return "negative" value
If thread_time_to_wakeup() was called with thread->sands in the
past compared to time_now, the function overflowed and returned an
extremely high value, resulting in the thread not being scheduled
for another 585,000 years or so (on 64 bit systems).

This commit makes thread_time_to_wakeup() return 1 (microsecond) if
thread->sands is earlier than time_now.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>