Fix building with --enable-debug option, but add deliberate error
Issue #1444 identified that keepalived would not build with the
--enable-debug option. This commits resolves the compilation errors
but also adds a #error statement if --enable-debug is selected, so
that the source code has to be editted to be able to build with the
option.

Since --enable-debug is not expected to work properly and is only
for debugging purposes, the #error statement will stop it being used
accidentally or being enabled in a distro's build of keepalived, for
example as Gentoo had done. Only developers/maintainers who are
prepared to edit the source code will be able to use --enable-debug.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>