Clear previous parser errors at startup or when reloading
When keepalived starts, the parent process reads the config. If there
was an error such as a missing } or extra { the skip_block flags was
left set, so when the child processes read the configuration, they
just skipped everything. This also meant that the configuration was
not correctly parser when reloading.

All parser state variables are now reinitialised before the configuration
is read.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>