Fix a checker process segfault when reloading with ha_suspend
If ha_suspend was enabled, when the checker process reloaded, it
left an old pointer to the read thread for the netlink monitoring
fd. When the process subsequently terminated, it tried cancelling
the thread using the old pointer, and would often segfault.

This also meant that after a reload, the checker process would no
longer be monitoring address changes, in other words ha_suspend
would not work.

Before a reload, both for the vrrp and the checker process, the
thread is cancelled and the pointer to the thread cleared. After
a reload the checker process now adds a new read thread, so ha_suspend
will continue working after the reload.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>