Add tools/timed_reload
This script makes keepalived reload its configuration at a specified time.
If there are several keepaliveds running on different systems, then can be
scheduled to all reload at the same time, by running the script on each system.
This is useful if the old and new configurations are not compatible with each
other.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>