Add option to set preferred lifetime for static and virtual IPv6 addresses
Generally  virtual ip addresses will want to be deprecated, but static
IPv6 addresses may or may not want to be deprecated. Previously the code
deprecated IPv6 addresses only if that had a /128 mask.

This commit retains the old settings as default, but now allows the
preferred_lft to be specified for each address.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>