Add dynamically incrementing process priority
If keepalived is delayed being scheduled after a timer expires, this
could cause delays in adverts being sent, or other issues. This would,
in practice, only occur on a very heavily loaded system.

This commit makes keepalived compare the time when epoll_wait() to
the time when it should have been scheduled due to the soonest
expiring timer. If keepalived is more than 1 second later being
scheduled to run, if will transition to realtime scheduling, if not
already using that, or if it is will increase the realtime priority
by 1 up to the system limit.

This approach should ensure that keepalived elevates itself to sufficient
priority so that it isn't delayed unduly in being scheduled.

In order to protect the system from keepalived, keepalived sets
RLIMIT_RTTIME to limits keepalived's overall CPU usage.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>