Don't use iptables if iptables command is iptables-nft
If the iptables command is iptables-nft, i.e. executing iptables -V
outputs something like:
  iptables v1.8.4 (nf_tables)
then using the iptables command will generate nftables configuration
and not iptables configuration. This means that the native iptables
chains that have to be set up in order for keepalived to be able to
work will not exist in the iptables configuration, but will be in
the nftables configuration. Consequently keepalived will fail to add
any rules to the iptables chains, since they don't exist.

The solution to this is that if the iptables command is iptables-nft,
then keepalived will use nft natively rather than using iptables.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>