Assign correct link local IPv6 address to macvlan moved namespace
When keepalived is configured to use a macvlan on a macvlan that
has been moved to a different namespace from its parent interface,
it needs to get the IPv6 link local address from the interface that
keepalived's macvlan is configured on. In the circumstance above,
the created macvlan's base_ifp cannot be set, so the vrrp's
configured_ifp must be used.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>