Set address family of virtual server from virtual server group
If all a virtual server's real servers are tunnelled, that gives no
indication of which address family the virtual server is intended for.
If the virtual server uses a virtual server group, and that vsg has
IPv4 or IPv6 addresses, these can be used to set the address family
of the virtual server.

If the virtual server group only has fwmarks, and all the real servers
are tunnelled, then default to IPv4.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>