Allow mixed IPv4/6 virtual server groups where all RS tunnelled
If all real and sorry servers of a virtual server are tunnelled, then
the address family of the virtual server can be either IPv4 or IPv6.
This means that it makes sense to have a virtual server group with both
IPv4 and IPv6 addresses, provided that all the virtual servers using it
meet the requirement of all real and sorry servers being tunnelled.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>