Add startup and shutdown scripts
This commit adds a feature to run a script when keepalived starts
up, and another script when keepalived shuts down.

The original motivation for this was that keepalived can setup IPVS
configuration that uses firewall marks, but it had no ability to
add the iptables/nftables configuration to set the fwmarks. Running
a script a startup allows the iptables/nftables configuration to be
added, and the shutdown script can remove the configuration.

This feature can also be used to setup the necessary iptables entries
if using vrrp_iptables, or to set interface configuration settings. In
fact, these can do anything that a script or program can do.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>