Don't set random fwd method if invalid lvs_method/lb_kind specified for VS
Issue #1550 identified that random forwarding methods were being set for
real servers. This was caused by specifying 'lb_kind nat' for the default
forwarding method on virtual servers. Since 'nat' is not valid (the
forwarding methods are specified in upper case (capital letters), a random
value was set for the forwarding method. This commit ensures that valid
forwarding methods are always set.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>