Fix intermittent segfault after config reload
The reload_{bfd,check,vrrp}_thread functions call thread_cleanup_master()
which frees all threads including the current thread. As a consequence, on
return to process_threads() after calling the reload_*_thread() function,
the thread variable is pointing to a thread structure that has been freed.

This commit makes process_threads() check if the current thread has been
freed, and if so updates the thread variable to NULL so that it won't be
used or queued to the thread_unuse queue.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>