Allow both master and backup IPVS sync daemons to run simultaneously
With load balancing there could be two of more systems handling
virtual servers, with either one able to take over from the other. In
that case, both systems need to be able to synchronize their IPVS
connections to the other, and this necessitates running both a master
and a backup sync daemon in each system. Furthermore, there is no need
to track a VRRP instance, since if all relevant VRRP instances are in
backup mode, no connections will be received by that system and so the
master sync daemon will not send and sync updates.

This commit allows lvs sync daemons to be configured without a VRRP
instance being specified, and in this case keepalived will set up
both a master and a backup sync daemon which will run for the
duration of keepalived running.

The above scenarion is not limited to having two systems handling
virtual servers, and in practice there is no practical limit to the
number of systems involved, but they must all use the same syncid and
multicast address.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>