Improve handling by parent process of children terminating
Issue #1561 identified that if a child process segfaulted while it was
terminating, the segfault was not reported by the parent process, making
the termination look as though nothig had failed.

This commit now ensures that if a child process does not exit cleanly then
the parent process will report it.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>