Fix handling SNMP file descriptors when epoll return read error
Issue #1570 identified that keepalived was segfaulting after reloading
when SNMP was being used. This commit now resolves the segfault.

In addition, this commit now only calls snmp_epoll_info() after an SNMP
timeout or activity on one of SNMP's file descriptors rather than before
every call of epoll_wait(). The basis for this is that the state of the
SNMP code can only change if it has been called, i.e. if snmp_read_thread()
or snmp_timeout_thread() is called.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>